src/main/java/ood/srp/report/ReportForProgrammer.java
package ood.srp.report;

import ood.srp.formatter.DateTimeParser;
import ood.srp.model.Employee;
import ood.srp.store.Store;

import java.util.Calendar;
import java.util.function.Predicate;

public class ReportForProgrammer implements Report {

    private final Store store;
    private final DateTimeParser<Calendar> dateTimeParser;

    public ReportForProgrammer(Store store, DateTimeParser<Calendar> dateTimeParser) {
        this.store = store;
        this.dateTimeParser = dateTimeParser;
    }

    @Override
    public String generate(Predicate<Employee> filter) {
        StringBuilder text = new StringBuilder();
        text.append("Name,Hired,Fired,Salary")
                .append(System.lineSeparator());
        for (Employee employee : store.findBy(filter)) {
            text.append(employee.getName()).append(",")
                    .append(dateTimeParser.parse(employee.getHired())).append(",")
                    .append(dateTimeParser.parse(employee.getFired())).append(",")
                    .append(employee.getSalary())
                    .append(System.lineSeparator());
        }
        return text.toString();
    }
}
